Add files via upload
Just introducing classifiers in the code. Next aim is to add multiple classifiers, compare them in a tabular format, display their accuracy in graphical representation and point out the best classifier according to the visualization.

Before that, number of neighbors in KNN will be changed to infer the optimum value of n_neighbors.